Fix addressing of spilled registers

+++ app/src/main/kotlin/compiler/lowlevel/Addressing.kt
@@ -4,7 +4,7 @@ import compiler.intermediate.Register
 
 sealed class Addressing {
     sealed class MemoryAddress {
-        data class Const(val address: ULong) : MemoryAddress() {
+        data class Const(val address: Int) : MemoryAddress() {
             override fun toString() = address.toString()
         }
         data class Label(val label: String) : MemoryAddress() {
@@ -22,10 +22,10 @@ sealed class Addressing {
 
     data class Base(
         val base: Register,
-        val displacement: MemoryAddress = MemoryAddress.Const(0U),
+        val displacement: MemoryAddress = MemoryAddress.Const(0),
     ) : Addressing() { // [base + displacement] or [base]
         override fun toAsm(registers: Map<Register, Register>) =
-            if (displacement == MemoryAddress.Const(0U)) {
+            if (displacement == MemoryAddress.Const(0)) {
                 "[${registers[base]!!.toAsm()}]"
             } else {
                 "[${registers[base]!!.toAsm()} + $displacement]"
@@ -36,11 +36,11 @@ sealed class Addressing {
         val base: Register,
         val index: Register,
         val scale: UByte, // either 1,2,4,8
-        val displacement: MemoryAddress = MemoryAddress.Const(0U)
+        val displacement: MemoryAddress = MemoryAddress.Const(0)
     ) : Addressing() { // [base + (index * scale) + displacement], [base + (index * scale)], [base + index + displacement], or [base + index]
         private val oneUByte: UByte = 1u
         override fun toAsm(registers: Map<Register, Register>) =
-            if (displacement == MemoryAddress.Const(0U)) {
+            if (displacement == MemoryAddress.Const(0)) {
                 if (scale == oneUByte) {
                     "[${registers[base]!!.toAsm()} + ${registers[index]!!.toAsm()}]"
                 } else {
@@ -58,10 +58,10 @@ sealed class Addressing {
     data class IndexAndDisplacement(
         val index: Register,
         val scale: UByte, // either 1,2,4,8
-        val displacement: MemoryAddress = MemoryAddress.Const(0U)
+        val displacement: MemoryAddress = MemoryAddress.Const(0)
     ) : Addressing() { // [(index*scale)] or [(index*scale) + displacement]
         override fun toAsm(registers: Map<Register, Register>) =
-            if (displacement == MemoryAddress.Const(0U)) {
+            if (displacement == MemoryAddress.Const(0)) {
                 "[(${registers[index]!!.toAsm()} * $scale)]"
             } else {
                 "[(${registers[index]!!.toAsm()} * $scale) + $displacement]"
